+++ module/combat.mjs
@@ -1,5 +1,6 @@
 import { HEROSYS } from "./herosystem6e.mjs";
-import { clamp, isGameV12OrLater } from "./utility/compatibility.mjs";
+import { clamp } from "./utility/compatibility.mjs";
+import { whisperUserTargetsForActor } from "./utility/util.mjs";
 
 export class HeroSystem6eCombat extends Combat {
     constructor(data, context) {
@@ -8,6 +9,26 @@ export class HeroSystem6eCombat extends Combat {
         this.previous = this.previous || {
             combatantId: null,
         };
+    }
+
+    getUniqueTokens() {
+        const results = [];
+        for (const c of this.combatants.values()) {
+            if (!results.find((o) => o.id === c.token.object.id)) {
+                results.push(c.token.object);
+            }
+        }
+        return results;
+    }
+
+    getUniqueCombatants() {
+        const results = [];
+        for (const c of this.combatants.values()) {
+            if (!results.find((o) => o.token.object.id === c.token.object.id)) {
+                results.push(c);
+            }
+        }
+        return results;
     }
 
     /**
@@ -20,39 +41,148 @@ export class HeroSystem6eCombat extends Combat {
      * @return {Promise<Combat>}        A promise which resolves to the updated Combat entity once updates are complete.
      */
 
-    async rollInitiative() {
-        //console.log("rollInitiative");
-        // Iterate over Combatants, performing an initiative roll for each
-        const updates = [];
-        for (let [id /*, value*/] of this.combatants.entries()) {
-            // Get Combatant data (non-strictly)
-            const combatant = this.combatants.get(id);
-            if (!combatant?.isOwner) return this;
-            //if (combatant.hasRolled) continue;
+    async rollInitiative(ids) {
+        if (CONFIG.debug.combat) {
+            console.debug(`Hero | rollInitiative`, ids);
+        }
+        // We need a unique combatant for each phase the actor acts.
 
-            if (!combatant.actor) continue;
+        for (const t of this.getUniqueTokens()) {
+            let tokenCombatants = this.combatants.filter(
+                (o) => o.tokenId === t.id && t.inCombat,
+            );
 
-            // Produce an initiative roll for the Combatant
-            let characteristic =
-                combatant.actor.system?.initiativeCharacteristic || "dex";
-            let dexValue =
-                combatant.actor.system.characteristics[characteristic].value;
-            let spdValue = combatant.actor.system.characteristics.spd.value;
-            //let intValue = combatant.actor.system.characteristics.int.value;
+            const lightningReflexes = t.actor?.items.find(
+                (o) =>
+                    o.system.XMLID === "LIGHTNING_REFLEXES_ALL" ||
+                    o.system.XMLID === "LIGHTNING_REFLEXES_SINGLE",
+            );
+            const lightningReflexesLevels = parseInt(
+                lightningReflexes?.system.LEVELS?.value ||
+                    lightningReflexes?.system.LEVELS ||
+                    lightningReflexes?.system.levels ||
+                    lightningReflexes?.system.other.levels ||
+                    0,
+            );
 
-            let initiativeValue = dexValue + spdValue / 100; // + intValue / 10000;
+            //Create extra combatants to match SPEED
+            let needToCreate =
+                parseInt(
+                    Math.max(1, t.actor?.system.characteristics.spd.value),
+                ) *
+                    (lightningReflexes ? 2 : 1) -
+                tokenCombatants.length;
 
-            if (initiativeValue != combatant.initiative) {
-                updates.push({
-                    _id: id,
-                    initiative: initiativeValue || 0,
+            const toCreate = [];
+            for (let i = 0; i < needToCreate; i++) {
+                toCreate.push({
+                    tokenId: t.id,
+                    sceneId: t.scene.id,
+                    actorId: t.actor.id,
+                    hidden: t.document.hidden,
+                });
+            }
+
+            if (toCreate.length > 0) {
+                const created = await this.createEmbeddedDocuments(
+                    "Combatant",
+                    toCreate,
+                );
+                tokenCombatants = tokenCombatants.concat(created);
+            }
+
+            // Perhaps we have too many
+            const toDelete = [];
+            for (let i = 0; i < -needToCreate; i++) {
+                toDelete.push(tokenCombatants[i].id);
+            }
+            if (toDelete.length > 0) {
+                await this.deleteEmbeddedDocuments("Combatant", toDelete);
+            }
+
+            // Produce an initiative roll for the Combatant.
+            const characteristic =
+                tokenCombatants[0].actor?.system?.initiativeCharacteristic ||
+                "dex";
+            const initValue =
+                tokenCombatants[0].actor?.system.characteristics[characteristic]
+                    .value || 0;
+            const spdValue =
+                tokenCombatants[0].actor?.system.characteristics.spd.value || 0;
+            const initiativeValue = initValue + spdValue / 100;
+
+            // Assign a segment and Initiative
+            let idx = 0;
+            for (let s = 1; s <= 12; s++) {
+                if (
+                    HeroSystem6eCombat.hasPhase(
+                        Math.max(
+                            1,
+                            tokenCombatants[idx].actor?.system.characteristics
+                                .spd.value || 0,
+                        ),
+                        s,
+                    )
+                ) {
+                    if (lightningReflexes) {
+                        const lightningReflexesAlias = `(${
+                            lightningReflexes.system.OPTION_ALIAS ||
+                            lightningReflexes.system.INPUT ||
+                            "All Actions"
+                        })`;
+                        if (
+                            tokenCombatants[idx].flags.segment !== s ||
+                            tokenCombatants[idx].flags.initiative !==
+                                initiativeValue + lightningReflexesLevels ||
+                            tokenCombatants[idx].flags
+                                .lightningReflexesAlias !==
+                                lightningReflexesAlias
+                        ) {
+                            await tokenCombatants[idx].update({
+                                "flags.segment": s,
+                                initiative:
+                                    initiativeValue + lightningReflexesLevels,
+                                "flags.lightningReflexesAlias":
+                                    lightningReflexesAlias,
+                            });
+                        }
+                        idx++;
+                    }
+
+                    if (
+                        tokenCombatants[idx].flags.segment !== s ||
+                        tokenCombatants[idx].flags.initiative !==
+                            initiativeValue
+                    ) {
+                        await tokenCombatants[idx].update({
+                            "flags.segment": s,
+                            initiative: initiativeValue,
+                            "-flags.lightningReflexesAlias": null,
+                        });
+                    }
+                    idx++;
+                }
+            }
+
+            // Rare case where SPD <= 0 (and for actorless tokens)
+            // NOTE: There is no code to prevent a SPD 0 token from acting, currently GM player needs to handle that manually.
+            // A SPD 0 character can't act, but does get a postSegment12.  In theory the SPD drain will eventually fade.
+            if (
+                (tokenCombatants[0].actor?.system.characteristics.spd.value ||
+                    0) <= 0 &&
+                tokenCombatants[0].flags.segment !== 12
+            ) {
+                tokenCombatants[0].flags.segment = 12;
+                await tokenCombatants[0].update({
+                    "flags.segment": 12,
+                    initiative: 10 + initiativeValue,
+                    "-flags.lightningReflexesAlias": null,
                 });
             }
         }
-        if (!updates.length) return this;
 
-        // Update multiple combatants
-        await this.updateEmbeddedDocuments("Combatant", updates);
+        // Likely new turn orders so setupTurns
+        this.setupTurns();
 
         return this;
     }
@@ -65,98 +195,24 @@ export class HeroSystem6eCombat extends Combat {
      */
 
     setupTurns() {
-        //console.log("setupTurns");
-        // Roll Initiative everytime as DEX/INT/SPD may have changed
-        // await this.rollAll();
+        if (CONFIG.debug.combat) {
+            console.debug(`Hero | Combat Setup Turns`);
+        }
+
+        // Likely never needed, but here for future reference.
+        // super.setupTurns();
+
         // Determine the turn order and the current turn
-        const turnsRaw = this.combatants.contents.sort(this._sortCombatants);
-
-        // Loop thru all combatants, add extra turn if they have LIGHTNING REFLEXES
-        for (const combatant of this.combatants) {
-            if (!combatant.actor) {
-                continue; // Not sure how this could happen
-            }
-
-            const lightningReflexes = combatant.actor.items.find(
-                (o) =>
-                    o.system.XMLID === "LIGHTNING_REFLEXES_ALL" ||
-                    o.system.XMLID === "LIGHTNING_REFLEXES_SINGLE",
-            );
-            if (lightningReflexes) {
-                const levels =
-                    lightningReflexes.system.LEVELS?.value ||
-                    lightningReflexes.system.LEVELS ||
-                    lightningReflexes.system.levels ||
-                    lightningReflexes.system.other.levels ||
-                    0;
-                const lightning_reflex_initiative = combatant.initiative
-                    ? combatant.initiative + parseInt(levels)
-                    : null;
-                const alias =
-                    lightningReflexes.system.OPTION_ALIAS ||
-                    lightningReflexes.system.INPUT ||
-                    "All Actions";
-                const lightning_reflex_alias = "(" + alias + ")";
-
-                const combatantLR = new Combatant({
-                    tokenId: combatant.tokenId,
-                    sceneId: combatant.sceneId,
-                    actorId: combatant.actor.id,
-                    hidden: combatant.hidden,
-                    _id: combatant.id,
-                });
-
-                combatantLR.initiative = lightning_reflex_initiative;
-                combatantLR.flags.lightningReflexesAlias =
-                    lightning_reflex_alias;
-
-                turnsRaw.push(combatantLR);
-
-                // Notice we didn't update the database with combatantLR.
-                // Not really sure how we would do that without really messing things up.
-                // As a result the flags.lightning_reflex_initiative is likely null,
-                // which we will update in the combatTracker::getData()
-            }
-        }
-
-        turnsRaw.sort(this._sortCombatants);
-
-        // this.turns is an array of combatants.  These combatants typically appear more than
-        // once in the array (1/SPEED).  Any change to one combatant for a specific token seems
-        // to update all combatant entries for that specific token.  So we can't store unique
-        // data in combatant (like segment number).  Using turnsExtra to store unique data (like segment).
-        //this.turnsExtra = []
-
-        // Assign Combatant copies to appropriate segments
-        // Notice segment[0] is unused
-        let turns = [];
-        for (let s = 1; s <= 12; s++) {
-            for (let t = 0; t < turnsRaw.length; t++) {
-                if (!turnsRaw[t].actor) {
-                    continue;
-                }
-                if (
-                    HeroSystem6eCombat.hasPhase(
-                        turnsRaw[t].actor.system.characteristics.spd.value,
-                        s,
-                    )
-                ) {
-                    const combatant = new Combatant(turnsRaw[t]);
-                    combatant.flags = {
-                        ...turnsRaw[t].flags,
-                        segment: s,
-                        turn: turns.length,
-                    };
-                    turns.push(combatant);
-                }
-            }
-        }
-
+        const turns = this.combatants.contents.sort(this._sortCombatants);
         if (this.turn !== null)
             this.turn = clamp(this.turn, 0, turns.length - 1);
 
-        // Update state tracking
-        let c = turns[this.turn];
+        // Update state tracking (v12)
+        // const c = turns[this.turn];
+        // this.current = this._getCurrentState(c);
+
+        // Update state tracking (v11)
+        const c = turns[this.turn];
         this.current = {
             round: this.round,
             turn: this.turn,
@@ -172,18 +228,25 @@ export class HeroSystem6eCombat extends Combat {
     }
 
     _sortCombatants(a, b) {
-        const initA = Number.isNumeric(a.initiative) ? a.initiative : -9999;
-        const initB = Number.isNumeric(b.initiative) ? b.initiative : -9999;
+        const segmentA = parseInt(a.flags.segment) || 12;
+        const segmentB = parseInt(b.flags.segment) || 12;
+        const initA = parseFloat(a.initiative) || 0;
+        const initB = parseFloat(b.initiative) || 0;
 
-        let initDifference = initB - initA;
-        if (initDifference != 0) {
+        let segmentDifference = segmentA - segmentB;
+        if (segmentDifference !== 0) {
+            return segmentDifference;
+        }
+
+        const initDifference = initB - initA;
+        if (initDifference !== 0) {
             return initDifference;
         }
 
         const typeA = a.actor?.hasPlayerOwner || a.hasPlayerOwner;
         const typeB = b.actor?.hasPlayerOwner || b.hasPlayerOwner;
 
-        if (typeA != typeB) {
+        if (typeA !== typeB) {
             if (typeA) {
                 return -1;
             }
@@ -191,32 +254,10 @@ export class HeroSystem6eCombat extends Combat {
                 return 1;
             }
         }
-    }
 
-    _sortSegments(a, b) {
-        const initA = Number.isNumeric(a.combatant.initiative)
-            ? a.combatant.initiative
-            : -9999;
-        const initB = Number.isNumeric(b.combatant.initiative)
-            ? b.combatant.initiative
-            : -9999;
-
-        let initDifference = initB - initA;
-        if (initDifference != 0) {
-            return initDifference;
-        }
-
-        const typeA = a.combatant.hasPlayerOwner;
-        const typeB = b.combatant.hasPlayerOwner;
-
-        if (typeA != typeB) {
-            if (typeA) {
-                return -1;
-            }
-            if (typeB) {
-                return 1;
-            }
-        }
+        // Force consistant sorting by token.id
+        //console.warn("Sorting undetermined. Using token.id to break the tie.");
+        return a.token.id.localeCompare(b.token.id);
     }
 
     // Standard HeroSystem rules per SPEED CHART
@@ -255,28 +296,6 @@ export class HeroSystem6eCombat extends Combat {
     /* -------------------------------------------- */
 
     /** @inheritdoc */
-    _onUpdate(data, options, userId) {
-        const priorState = foundry.utils.deepClone(this.current);
-        super._onUpdate(data, options, userId);
-
-        // KLUGE: FoundryVTT V12 added a stateChanged check before calling manageTurnEvents that breaks our system.
-        // The default #recordPreviousState looks for the same CombatantID, which we typically have several.
-        // We use a Hero6e version #recordPreviousState.
-        if (isGameV12OrLater()) {
-            const stateChanged = this.#recordPreviousState(priorState);
-            if (stateChanged && options.turnEvents !== false)
-                this._manageTurnEvents();
-        }
-
-        // _onUpdate isn't async, so can't call await.
-        // Without await is seems to loose track of turn (go from turn=0 to turn=last ).
-        // Instead moved scrollToTurn to combatTracker::_render.
-        //await ui.combat.scrollToTurn();
-    }
-
-    /* -------------------------------------------- */
-
-    /** @inheritdoc */
     async _onCreateDescendantDocuments(
         parent,
         collection,
@@ -285,7 +304,9 @@ export class HeroSystem6eCombat extends Combat {
         options,
         userId,
     ) {
-        //console.log("_onCreateDescendantDocuments");
+        if (CONFIG.debug.combat) {
+            console.debug(`Hero | _onCreateDescendantDocuments`, this);
+        }
 
         //Missing actor?
         // let missingActors = documents.filter((o) => !o.actor);
@@ -303,16 +324,6 @@ export class HeroSystem6eCombat extends Combat {
         // Get current combatant
         const oldCombatant = this.combatant;
 
-        // Super
-        await super._onCreateDescendantDocuments(
-            parent,
-            collection,
-            documents,
-            data,
-            options,
-            userId,
-        );
-
         // Setup turns in segment fashion
         //const _turns = this.setupTurns();
 
@@ -329,8 +340,18 @@ export class HeroSystem6eCombat extends Combat {
             await this.update({ turn: this.turn });
         }
 
+        // Super
+        await super._onCreateDescendantDocuments(
+            parent,
+            collection,
+            documents,
+            data,
+            options,
+            userId,
+        );
+
         // Render the collection
-        if (this.active) this.collection.render();
+        //if (this.active) this.collection.render();
     }
 
     /* -------------------------------------------- */
@@ -354,7 +375,7 @@ export class HeroSystem6eCombat extends Combat {
             this.turns[this.turn + 1 > this.turns.length ? 0 : this.turn + 1];
 
         // Super
-        super._onDeleteDescendantDocuments(
+        await super._onDeleteDescendantDocuments(
             parent,
             collection,
             documents,
@@ -362,6 +383,21 @@ export class HeroSystem6eCombat extends Combat {
             options,
             userId,
         );
+
+        // Make sure we delete all combatants with the same tokenID
+        if (collection === "combatants") {
+            for (const doc of documents) {
+                const toDelete = this.combatants.filter(
+                    (c) => c.token.id === doc.token.id,
+                );
+                if (toDelete.length > 0) {
+                    await this.deleteEmbeddedDocuments(
+                        "Combatant",
+                        toDelete.map((c) => c.id),
+                    );
+                }
+            }
+        }
 
         // Setup turns in segment fashion
         // this.setupTurns();
@@ -394,6 +430,8 @@ export class HeroSystem6eCombat extends Combat {
             }
         }
         await this.update({ turn: this.turn, round: this.round });
+
+        await this.rollInitiative();
 
         // Render the collection
         if (this.active) this.collection.render();
@@ -511,7 +549,7 @@ export class HeroSystem6eCombat extends Combat {
                 );
                 if (value - spentEnd >= end) {
                     spentEnd += end;
-                    if (end > 0) {
+                    if (end >= 0) {
                         content += `<li>${powerUsingEnd.name} (${end})</li>`;
                     }
                 } else {
@@ -533,7 +571,7 @@ export class HeroSystem6eCombat extends Combat {
             }
         }
 
-        if (content != "" && !this.combatant.isFake) {
+        if (content != "" && !this.combatant.isFake && spentEnd > 0) {
             let segment = this.combatant.flags.segment;
             let value = parseInt(
                 this.combatant.actor.system.characteristics.end.value,
@@ -553,11 +591,12 @@ export class HeroSystem6eCombat extends Combat {
                 token,
             });
             speaker["alias"] = combatant.actor.name;
+
             const chatData = {
                 user: game.user._id,
                 type: CONST.CHAT_MESSAGE_TYPES.OTHER,
                 content: content,
-                whisper: ChatMessage.getWhisperRecipients("GM"),
+                whisper: whisperUserTargetsForActor(combatant.actor),
                 speaker,
             };
 
@@ -670,7 +709,9 @@ export class HeroSystem6eCombat extends Combat {
         content += "<ul>";
         contentHidden += "<ul>";
         let hasHidden = false;
-        for (let combatant of this.combatants.filter((o) => !o.defeated)) {
+        for (const combatant of this.getUniqueCombatants().filter(
+            (o) => !o.defeated,
+        )) {
             const actor = combatant.actor;
 
             // Make sure we have a valid actor